Implement asymmetric encryption and update to v0.0.3
This patch introduces the asymmetric encryption/decryption approach
using the mincrypt public and private keys. The keys can be generated
using the --key-size and --key-file arguments passed to the mincrypt
binary and they are based on the password and salt values with added
randomicity so the recipient will have to have the private key as
well as the knowledge of the right salt and password values since
this all alters the initialization vectors used for encryption or
decryption purposes.

Also, the --dump-vectors argument and mincrypt_dump_vectors() function
has been added to dump the initialization vectors for debugging purposes
if debugging is required.

Michal

Signed-off-by: Michal Novotny <mignov@gmail.com>